requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.9.3
